rag_service/requirements.txt
flask==2.0.1
langchain==0.0.267
chromadb==0.4.6
openai==0.27.8
python-dotenv==0.19.0
pypdf==3.12.1 